feat: update the TinyDemo Nada program

+++ packages/nillion/next-project-programs/src/tinydemo.py
@@ -5,6 +5,6 @@ def nada_main():
     api_num = SecretInteger(Input(name="api_num", party=party1))
     guess = SecretInteger(Input(name="guess", party=party1))
 
-    new_int = (api_num >= guess).if_else((api_num <= guess).if_else(Integer(1), Integer(0)), Integer(0))
+    new_int = (api_num >= guess).if_else((api_num <= guess).if_else(Integer(1), Integer(2)), Integer(3))
 
     return [Output(new_int, "my_output", party1)]
